# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AssetKind": {
        "enum": [
          "image",
          "video",
          "audio"
        ],
        "title": "AssetKind",
        "type": "string"
      },
      "AssetLinkCreate": {
        "properties": {
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "position": {
            "default": 0,
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Position",
            "type": "integer"
          },
          "role": {
            "$ref": "#/components/schemas/AssetRole"
          }
        },
        "required": [
          "asset_id",
          "role"
        ],
        "title": "AssetLinkCreate",
        "type": "object"
      },
      "AssetLinkRead": {
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/AssetRead"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "role": {
            "$ref": "#/components/schemas/AssetRole"
          }
        },
        "required": [
          "role",
          "position",
          "asset"
        ],
        "title": "AssetLinkRead",
        "type": "object"
      },
      "AssetRead": {
        "properties": {
          "ai_generated": {
            "title": "Ai Generated",
            "type": "boolean"
          },
          "checksum_sha256": {
            "title": "Checksum Sha256",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/AssetKind"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "mime": {
            "title": "Mime",
            "type": "string"
          },
          "original_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Filename"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/AssetSource"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at",
          "kind",
          "source",
          "mime",
          "size_bytes",
          "checksum_sha256",
          "width",
          "height",
          "duration_s",
          "original_filename",
          "ai_generated",
          "metadata"
        ],
        "title": "AssetRead",
        "type": "object"
      },
      "AssetRole": {
        "enum": [
          "reference",
          "canonical",
          "front",
          "three_quarter",
          "full_body",
          "product_photo",
          "cutout",
          "mask",
          "photo",
          "footage",
          "drone",
          "floor_plan"
        ],
        "title": "AssetRole",
        "type": "string"
      },
      "AssetSource": {
        "enum": [
          "upload",
          "generated",
          "derived"
        ],
        "title": "AssetSource",
        "type": "string"
      },
...
      "Body_upload_asset_api_assets_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "JPEG, PNG, WEBP, MP4, MOV, WAV, MP3 or M4A",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_asset_api_assets_post",
        "type": "object"
      },
...
      "CharacterRead": {
        "properties": {
          "adult_age": {
            "title": "Adult Age",
            "type": "integer"
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetLinkRead"
            },
            "title": "Assets",
            "type": "array"
          },
          "body_description": {
            "title": "Body Description",
            "type": "string"
          },
          "brand_tone": {
            "title": "Brand Tone",
            "type": "string"
          },
          "clothing_preferences": {
            "title": "Clothing Preferences",
            "type": "string"
          },
          "color_palette": {
            "items": {
              "type": "string"
            },
            "title": "Color Palette",
            "type": "array"
          },
          "consent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_language": {
            "title": "Default Language",
            "type": "string"
          },
          "default_prompt": {
            "title": "Default Prompt",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "eye_color": {
            "title": "Eye Color",
            "type": "string"
          },
          "face_description": {
            "title": "Face Description",
            "type": "string"
          },
          "hair": {
            "title": "Hair",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_real_person": {
            "title": "Is Real Person",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "negative_prompt": {
            "title": "Negative Prompt",
            "type": "string"
          },
          "personality": {
            "title": "Personality",
            "type": "string"
          },
          "preferred_camera_angles": {
            "items": {
              "type": "string"
            },
            "title": "Preferred Camera Angles",
            "type": "array"
          },
          "presentation": {
            "title": "Presentation",
            "type": "string"
          },
          "skin_appearance": {
            "title": "Skin Appearance",
            "type": "string"
          },
          "speaking_style": {
            "title": "Speaking Style",
            "type": "string"
          },
          "style": {
            "title": "Style",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "voice_profile": {
            "additionalProperties": true,
            "title": "Voice Profile",
            "type": "object"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at",
          "name",
          "adult_age",
          "description",
          "presentation",
          "face_description",
          "hair",
          "eye_color",
          "skin_appearance",
          "body_description",
          "style",
          "clothing_preferences",
          "personality",
          "speaking_style",
          "brand_tone",
          "default_language",
          "default_prompt",
          "negative_prompt",
          "preferred_camera_angles",
          "color_palette",
          "voice_profile",
          "is_real_person",
          "consent_id",
          "assets"
        ],
        "title": "CharacterRead",
        "type": "object"
      },
...
      "ConsentCreate": {
        "properties": {
          "confirm": {
            "const": true,
            "description": "Must be true: confirms the consent statement.",
            "title": "Confirm",
            "type": "boolean"
          },
          "evidence_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Asset Id"
          },
          "granted_by": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Granted By",
            "type": "string"
          },
          "subject_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Subject Name",
            "type": "string"
          },
          "subject_type": {
            "$ref": "#/components/schemas/ConsentSubject"
          }
        },
        "required": [
          "subject_type",
          "subject_name",
          "granted_by",
          "confirm"
        ],
        "title": "ConsentCreate",
        "type": "object"
      },
      "ConsentRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Asset Id"
          },
          "granted_at": {
            "format": "date-time",
            "title": "Granted At",
            "type": "string"
          },
          "granted_by": {
            "title": "Granted By",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "statement": {
            "title": "Statement",
            "type": "string"
          },
          "subject_name": {
            "title": "Subject Name",
            "type": "string"
          },
          "subject_type": {
            "$ref": "#/components/schemas/ConsentSubject"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at",
          "subject_type",
          "subject_name",
          "statement",
          "granted_by",
          "granted_at",
          "evidence_asset_id",
          "revoked_at",
          "is_active"
        ],
        "title": "ConsentRead",
        "type": "object"
      },
      "ConsentSubject": {
        "enum": [
          "face",
          "voice"
        ],
        "title": "ConsentSubject",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_ConsentRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConsentRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ConsentRead]",
        "type": "object"
      },
...
      "ProductRead": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetLinkRead"
            },
            "title": "Assets",
            "type": "array"
          },
          "brand": {
            "title": "Brand",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cta": {
            "title": "Cta",
            "type": "string"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "title": "Features",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at",
          "name",
          "brand",
          "description",
          "features",
          "price",
          "currency",
          "cta",
          "website",
          "assets"
        ],
        "title": "ProductRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AI Influencer Studio",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/assets": {
      "post": {
        "operationId": "upload_asset_api_assets_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_asset_api_assets_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Asset",
        "tags": [
          "assets"
        ]
      }
    },
    "/api/assets/{asset_id}": {
      "delete": {
        "operationId": "delete_asset_api_assets__asset_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Asset",
        "tags": [
          "assets"
        ]
      },
      "get": {
        "operationId": "get_asset_api_assets__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Asset",
        "tags": [
          "assets"
        ]
      }
    },
    "/api/assets/{asset_id}/content": {
      "get": {
        "operationId": "get_asset_content_api_assets__asset_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Asset Content",
        "tags": [
          "assets"
        ]
      }
    },
...
    "/api/characters/{character_id}/assets": {
      "post": {
        "operationId": "attach_asset_api_characters__character_id__assets_post",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetLinkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Attach Asset",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/characters/{character_id}/assets/{asset_id}": {
      "delete": {
        "operationId": "detach_asset_api_characters__character_id__assets__asset_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detach Asset",
        "tags": [
          "characters"
        ]
      }
    },
    "/api/consents": {
      "get": {
        "operationId": "list_consents_api_consents_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ConsentRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Consents",
        "tags": [
          "consents"
        ]
      },
      "post": {
        "operationId": "create_consent_api_consents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Consent",
        "tags": [
          "consents"
        ]
      }
    },
    "/api/consents/{consent_id}": {
      "get": {
        "operationId": "get_consent_api_consents__consent_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "consent_id",
            "required": true,
            "schema": {
              "title": "Consent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Consent",
        "tags": [
          "consents"
        ]
      }
    },
    "/api/consents/{consent_id}/revoke": {
      "post": {
        "operationId": "revoke_consent_api_consents__consent_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "consent_id",
            "required": true,
            "schema": {
              "title": "Consent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Consent",
        "tags": [
          "consents"
        ]
      }
    },
...
    "/api/products/{product_id}/assets": {
      "post": {
        "operationId": "attach_asset_api_products__product_id__assets_post",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetLinkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Attach Asset",
        "tags": [
          "products"
        ]
      }
    },
    "/api/products/{product_id}/assets/{asset_id}": {
      "delete": {
        "operationId": "detach_asset_api_products__product_id__assets__asset_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detach Asset",
        "tags": [
          "products"
        ]
      }
    },
...
    "/api/properties/{property_id}/assets": {
      "post": {
        "operationId": "attach_asset_api_properties__property_id__assets_post",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "title": "Property Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetLinkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Attach Asset",
        "tags": [
          "properties"
        ]
      }
    },
    "/api/properties/{property_id}/assets/{asset_id}": {
      "delete": {
        "operationId": "detach_asset_api_properties__property_id__assets__asset_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "title": "Property Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detach Asset",
        "tags": [
          "properties"
        ]
      }
    },
...
  }
}